seperate index file! and staticing things, also renamming shooter
