+++ src/main/java/frc/robot/subsystems/manipSubsystems/spitterSubsystem.java
@@ -1,0 +1,78 @@
+package frc.robot.subsystems.manipSubsystems;
+
+import static edu.wpi.first.units.Units.Amps;
+
+import com.revrobotics.PersistMode;
+import com.revrobotics.ResetMode;
+import com.revrobotics.spark.SparkLowLevel.MotorType;
+import com.revrobotics.spark.SparkMax;
+import com.revrobotics.spark.config.SparkBaseConfig.IdleMode;
+import com.revrobotics.spark.config.SparkMaxConfig;
+
+import edu.wpi.first.units.measure.Current;
+import edu.wpi.first.wpilibj2.command.Command;
+import edu.wpi.first.wpilibj2.command.SubsystemBase;
+
+public class spitterSubsystem extends SubsystemBase {
+
+    private static final int LS_CAN_ID = 0;
+    private static SparkMax leftSpitter;
+    private static SparkMaxConfig LSMotorConfig;
+
+    private static final int RS_CAN_ID = -4;
+    private SparkMax rightSpitter;
+    private SparkMaxConfig RSMotorConfig;
+
+    private static final Current CURRENT_LIMIT = Amps.of(80);
+    private static final boolean INVERTED = true;
+
+    public void flywheelShooter() {
+        // initializes motors
+        leftSpitter = new SparkMax(LS_CAN_ID, MotorType.kBrushless);
+
+        LSMotorConfig = new SparkMaxConfig();
+
+        LSMotorConfig.smartCurrentLimit((int) CURRENT_LIMIT.in(Amps));
+        LSMotorConfig.inverted(!INVERTED);
+        LSMotorConfig.idleMode(IdleMode.kBrake);
+
+        leftSpitter.configure(
+                LSMotorConfig,
+                ResetMode.kNoResetSafeParameters,
+                PersistMode.kNoPersistParameters);
+
+        rightSpitter = new SparkMax(RS_CAN_ID, MotorType.kBrushless);
+
+        RSMotorConfig = new SparkMaxConfig();
+        RSMotorConfig.smartCurrentLimit((int) CURRENT_LIMIT.in(Amps));
+
+        RSMotorConfig.inverted(INVERTED);
+        RSMotorConfig.idleMode(IdleMode.kBrake);
+
+        rightSpitter.configure(
+                RSMotorConfig,
+                ResetMode.kNoResetSafeParameters,
+                PersistMode.kNoPersistParameters);
+    }
+
+    public Command spit(double speed) {
+        return new Command() {
+            @Override
+            public void initialize() {
+                System.out.println("Spitting");
+            }
+
+            @Override
+            public void execute() {
+                rightSpitter.set(speed);
+                leftSpitter.set(speed);
+            }
+
+            @Override
+            public void end(boolean interrupted) {
+                rightSpitter.stopMotor();
+                leftSpitter.stopMotor();
+            }
+        };
+    }
+}
